refactor: Add setter method in response class, Change dto name
