+++ back/src/main/java/com/github/oneline/onelinecourse/controller/comment/CommentResponse.java
@@ -8,6 +8,7 @@ import org.springframework.beans.BeanUtils;
 import java.time.LocalDateTime;
 
 @Getter
+@Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
