+++ back/src/main/java/com/github/oneline/onelinecourse/controller/user/UserController.java
@@ -22,10 +22,10 @@ public class UserController {
     // 회원 조회
     @GetMapping("/{userId}")    // GET 요청 방식의 API
     @ApiOperation(value = "유저 조회")
-    public ResponseEntity<ResponseUserDto> searchUser(
+    public ResponseEntity<UserResponse> searchUser(
             @PathVariable @ApiParam(value = "유저 ID", example = "1df43fh1246") String userId) {
         return ResponseEntity.ok(
-                new ResponseUserDto(userService.searchUser(userId))
+                new UserResponse(userService.searchUser(userId))
         );
     }
 
@@ -34,20 +34,20 @@ public class UserController {
     @ApiOperation(value = "최초 로그인 여부 확인 및 유저 등록")
     // ResponseEntity
     // HTTP 요청(Request) 또는 응답(Response)에 해당하는 HttpHeader와 HttpBody를 포함하는 클래스
-    public ResponseEntity<ResponseUserDto> createUser(
-            @RequestBody CreateUserRequestDto createUserRequestDto) { // @RequestBody : 클라이언트가 전송하는 Http 요청의 Body내용을 Java Object로 변환시켜주는 역할
+    public ResponseEntity<UserResponse> createUser(
+            @RequestBody UserCreateRequest createUserRequestDto) { // @RequestBody : 클라이언트가 전송하는 Http 요청의 Body내용을 Java Object로 변환시켜주는 역할
 
         // 회원이 등록되어 있는지 체크
         User user = userService.searchUser(createUserRequestDto.toEntity().getId());
         if(user != null) {  // 등록되어 있다면
             return ResponseEntity.ok(
-                    new ResponseUserDto(userService.searchUser(createUserRequestDto.toEntity().getId()))
+                    new UserResponse(userService.searchUser(createUserRequestDto.toEntity().getId()))
             );
         }
         // 등록되어 있지 않다면(db에 저장)
         // 응답 헤더의 상태 코드 본문을 직접 다루기 위해 사용
         return ResponseEntity.ok(   // 200 ok 상태코드
-                new ResponseUserDto(userService.createUser(createUserRequestDto.toEntity()))
+                new UserResponse(userService.createUser(createUserRequestDto.toEntity()))
         );
 
     }
